# backend/openapi.yaml
openapi: 3.0.0
info:
  title: GamePlatform API
  version: v1.0.0
  description: API for managing a gaming platform including championships, events, user registrations, results, and more.
servers:
  - url: /api # Assuming all backend routes are prefixed with /api
components:
  schemas:
...
    Registration:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: string
          format: uuid
        championship_id:
          type: integer
          nullable: true
        event_id:
          type: integer
          nullable: true
        team_id:
          type: integer
          nullable: true
        vehicle_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, waitlisted]
        registration_date:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        # Joined data (user, championship, event, team, vehicle) can be described here if always present
        # Example for user:
        # user: 
        #   $ref: '#/components/schemas/UserProfileShort' # A shorter version of UserProfile for embedding
      required:
        - user_id
        - status
    RegistrationCreation:
      type: object
      properties:
        championship_id:
          type: integer
          nullable: true
          description: Required if event_id is not provided.
        event_id:
          type: integer
          nullable: true
          description: Required if championship_id is not provided.
        team_id:
          type: integer
          nullable: true
        vehicle_id:
          type: integer
          nullable: true
      # Logic: either championship_id or event_id must be provided (handled in description/backend)
    RegistrationStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, waitlisted]
      required:
        - status

    # --- File Schemas ---
...
    UserProfileShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        avatar_url:
          type: string
          format: url
          nullable: true
  securitySchemes:
    bearerAuth: # Can be named anything
      type: http
      scheme: bearer
      bearerFormat: JWT # Supabase uses JWT
security:
  - bearerAuth: [] # Applies Bearer token auth globally to all operations by default. Override locally if needed.
paths:
...
  /registrations:
    post:
      summary: Register for an event or championship
      tags: [Registrations]
      security:
        - bearerAuth: [] # Requires authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCreation'
      responses:
        '201':
          description: Registration successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          description: Invalid input (e.g., missing IDs, both IDs provided).
        '401':
          description: Unauthorized.
        '409':
          description: Conflict (e.g., already registered).
    get:
      summary: Get registrations (Admin gets all with filters, User gets own)
      tags: [Registrations]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: user_id # Admin filter
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: event_id # Admin filter
          in: query
          required: false
          schema: { type: integer }
        - name: championship_id # Admin filter
          in: query
          required: false
          schema: { type: integer }
        - name: status # Admin filter
          in: query
          required: false
          schema: { type: string, enum: [pending, confirmed, cancelled, waitlisted] }
      responses:
        '200':
          description: A list of registrations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  registrations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Registration'
                  total: { type: integer }
                  # ... pagination fields
  /registrations/{id}:
    get:
      summary: Get a single registration by ID
      tags: [Registrations]
      security:
        - bearerAuth: [] # Admin or owner
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Registration details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '403':
          description: Forbidden (not owner or admin).
        '404':
          description: Registration not found.
  /registrations/{id}/status:
    put:
      summary: Update registration status
      tags: [Registrations]
      security:
        - bearerAuth: [] # Admin or owner (for cancellation)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationStatusUpdate'
      responses:
        '200':
          description: Status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          description: Invalid status or request.
        '403':
          description: Forbidden (user cannot perform this status update).
        '404':
          description: Registration not found.
          
  # --- File Routes ---
...